feat: transcription to typescript
