+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "chatgpt-io",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "ChatGPT Client API, Blazing Fast, without using browser",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc"
   },
   "repository": {
     "type": "git",
@@ -46,11 +47,12 @@
   "author": "PawanOsman",
   "license": "MIT",
   "dependencies": {
-    "socket.io-client": "^4.5.4",
-    "uuid": "^9.0.0"
+    "socket.io-client": "^4.5.4"
   },
   "devDependencies": {
-    "fastify": "^4.11.0"
+    "@types/node": "^18.11.18",
+    "fastify": "^4.11.0",
+    "typescript": "^4.9.4"
   },
   "bugs": {
     "url": "https://github.com/PawanOsman/chatgpt-io/issues"
